moved zNormalisation to utilites and cluster variales to abtract classes.
