+++ src/main/java/timeseriesweka/clusterers/AbstractTimeSeriesClusterer.java
@@ -9,6 +9,8 @@ import weka.clusterers.AbstractClusterer;
 import weka.core.Instance;
 import weka.core.Instances;
 
+import java.util.ArrayList;
+
 /**
  *
  * @author pfm15hbu
@@ -16,42 +18,19 @@ import weka.core.Instances;
 public abstract class AbstractTimeSeriesClusterer extends AbstractClusterer{
     
     protected boolean dontCopyInstances = false;
+
+    protected int[] cluster;
+    protected ArrayList<Integer>[] clusters;
+
+    public int[] getCluster(){
+        return cluster;
+    }
+
+    public ArrayList<Integer>[] getClusters(){
+        return clusters;
+    }
     
     public void setDontCopyInstances(boolean b){
         dontCopyInstances = b;
-    }
-
-    protected void zNormalise(Instances data) {
-        for (Instance inst: data){
-            zNormalise(inst);
-        }
-    }
-
-    protected void zNormalise(Instance inst){
-        double meanSum = 0;
-        int length = inst.numAttributes();
-
-        for (int i = 0; i < length; i++){
-            meanSum += inst.value(i);
-        }
-
-        double mean = meanSum / length;
-
-        double squareSum = 0;
-
-        for (int i = 0; i < length; i++){
-            double temp = inst.value(i) - mean;
-            squareSum += temp * temp;
-        }
-
-        double stdev = Math.sqrt(squareSum/(length-1));
-
-        if (stdev == 0){
-            stdev = 1;
-        }
-
-        for (int i = 0; i < length; i++){
-            inst.setValue(i, (inst.value(i) - mean) / stdev);
-        }
     }
 }
